// Lab 10/Solutions/client/apiclient/model/chat.ts
/**
 * Messenger API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Message } from './message';


export interface Chat { 
    /**
     * is group
     */
    group?: boolean;
    /**
     * user id or group id
     */
    id?: number;
    message?: Message;
    /**
     * user name or group name
     */
    name?: string;
    /**
     * true means online, false means offline
     */
    status?: boolean;
    /**
     * number of unread messages
     */
    unread?: number;
}